* CDRIVER-887-PRId64-correct-cast:
  CDRIVER-887: Incorrect cast for PRId64

+++ src/mongoc/mongoc-stream.c
@@ -489,8 +489,8 @@ _mongoc_stream_writev_full (mongoc_stream_t *stream,
 
    if (r != total_bytes) {
       bson_set_error (error, MONGOC_ERROR_STREAM, MONGOC_ERROR_STREAM_SOCKET,
-                      "Failure to send all requested bytes (only sent: %" PRId64 "/%" PRId64 " in %dms) during socket delivery",
-                      (long long)r, (unsigned long long)total_bytes, timeout_msec);
+                      "Failure to send all requested bytes (only sent: %" PRIu64 "/%" PRId64 " in %dms) during socket delivery",
+                      (uint64_t) r, (int64_t)total_bytes, timeout_msec);
 
       RETURN(false);
    }
